producer/requirements.txt
fastavro==1.8.4
requests==2.31.0
pytz==2023.3
hydra-core==1.3.2
mysql-connector-python==8.1.0
pyyaml==6.0.1
confluent_kafka==2.0.2
retrying==1.3.4

# sqlalchemy==2.0.21
# pymysql==1.1.0
# pandas==2.0.3